Python: mark Flexx and REMI as widget-centric
I don't feel like removing them, so let's loosen the criteria a bit.
With things like server-side React components whether something is
"pure DOM" or based on widgets is a slippery distinction anyway.  The
ability to define widgets in the language of the framework without
diving into separate HTML and JS templates seems like a good
requirement to settle on.